third_party/WebKit/LayoutTests/fast/scroll-behavior/snaps-after-scrollbar-dragging.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
body {
  margin: 0px;
}
div {
  position: absolute;
}
#scroller {
  width: 400px;
  height: 400px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
  padding: 0px;
}
.snap {
  width: 200px;
  height: 200px;
  background-color: blue;
  scroll-snap-align: start;
}
#space {
  width: 1000px;
  height: 1000px;
}
#left-top {
  left: 0px;
  top: 0px;
}
#right-top {
  left: 400px;
  top: 0px;
}
#left-bottom {
  left: 0px;
  top: 400px;
}

</style>

<div id='scroller'>
  <div id="space"></div>
  <div class="snap" id="left-top"></div>
  <div class="snap" id="right-top"></div>
  <div class="snap" id="left-bottom"></div>
</div>

<script>
var scroller = document.getElementById("scroller");
function waitForAnimationEnd() {
  const MAX_FRAME = 500;
  var last_changed_frame = 0;
  var last_scroller_x = scroller.scrollLeft;
  var last_scroller_y = scroller.scrollTop;
  return new Promise((resolve, reject) => {
    function tick(frames) {
      // We requestAnimationFrame either for 500 frames or until 20 frames with
      // no change have been observed.
      if (frames >= MAX_FRAME || frames - last_changed_frame > 20) {
        resolve();
      } else {
        if (scroller.scrollLeft != last_scroller_x ||
            scroller.scrollTop != last_scroller_y) {
          last_changed_frame = frames;
          last_scroller_x = scroller.scrollLeft;
          last_scroller_y = scroller.scrollTop;
        }
        requestAnimationFrame(tick.bind(this, frames + 1));
      }
    }
    tick(0);
  });
}

function dragScrollbar(x, y, w, h) {
  return new Promise((resolve, reject) => {
    chrome.gpuBenchmarking.smoothDrag(x, y, x + w, y + h, resolve, 2);
  });
}

function vertical_scrollbar() {
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollTop, 0);
  assert_equals(scroller.scrollLeft, 0);
  return dragScrollbar(398, 20, 0, 100)
  .then(waitForAnimationEnd)
  .then(() => {
    assert_equals(scroller.scrollTop, 400);
  })
}

function horizontal_scrollbar() {
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollTop, 0);
  assert_equals(scroller.scrollLeft, 0);
  return dragScrollbar(20, 398, 100, 0)
  .then(waitForAnimationEnd)
  .then(() => {
    assert_equals(scroller.scrollLeft, 400);
  })
}

promise_test(t => {
  return vertical_scrollbar();
}, "Snaps after dragging the vertical scrollbar.");

promise_test(t => {
  return horizontal_scrollbar();
}, "Snaps after dragging the horizontal scrollbar.");

</script>
